requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
docopt==0.6.2
Flask==1.0.2
idna==2.7
itchat==1.3.10
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
multi-key-dict==2.0.3
numpy==1.15.4
pbr==5.1.1
Pillow==5.3.0
pipreqs==0.4.9
pyparsing==2.3.0
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.5
python-jenkins==1.4.0
requests==2.20.1
six==1.11.0
torch==0.4.1
torchvision==0.2.1
urllib3==1.24.1
Werkzeug==0.14.1
wincertstore==0.2
yarg==0.1.9
